perlin noise

+++ utils.hpp
@@ -16,7 +16,7 @@ inline double random_double(double min = 0, double max = 1) {
 	return distribution(generator);
 }
 
-inline double random_int(int min, int max) {
+inline int random_int(int min, int max) {
 	return static_cast<int>(random_double(min, max + 1));
 }
